parallel upload

+++ git-remote-gosh/src/blockchain/commit/save.rs
@@ -68,7 +68,7 @@ pub async fn notify_commit(
     context: &mut GitHelper,
     commit_id: &ObjectId,
     branch: &str,
-    number_of_files_changed: u64
+    number_of_files_changed: u32
 ) -> Result<(), Box<dyn Error>> {
     let wallet = user_wallet(context)?;
     let params = serde_json::json!({
